+++ TestProject/app/src/main/java/net/aquadc/mike/plugin/test/BigIntegersKt.kt
@@ -8,5 +8,11 @@ fun whatever() {
     println(BigInteger.valueOf(0))
     println(BigDecimal("1"))
     println(BigInteger("1"))
-    println(BigDecimal("1").compareTo(BigDecimal.ZERO))
+    println(BigDecimal.ONE.compareTo(BigDecimal.ZERO))
+    println(BigInteger.ONE.compareTo(BigInteger.ZERO))
+    println(BigInteger.ONE.let { it.compareTo(BigInteger.ZERO) })
+    println(BigInteger.ONE.run {
+        println()
+        compareTo(BigInteger.ZERO)
+    })
 }