+++ src/main/kotlin/net/aquadc/mike/plugin/bigdecimal/BigDecimalConstantInspection.kt
@@ -12,7 +12,6 @@ import net.aquadc.mike.plugin.resolvedClassFqn
 import net.aquadc.mike.plugin.test
 import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
 import org.jetbrains.kotlin.psi.KtElement
-import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
 import org.jetbrains.uast.UCallExpression
 import org.jetbrains.uast.UExpression
 import org.jetbrains.uast.UastCallKind.Companion.CONSTRUCTOR_CALL
@@ -65,10 +64,8 @@ class BigDecimalConstantInspection : UastInspection(), CleanupLocalInspectionToo
         fun visitNewExpression(expression: UCallExpression) { // TODO gets triggered several times for single expression
             val src = expression.sourcePsi ?: return
             val className = expression.resolvedClassFqn ?: return
-            val argVal = expression
-                .takeIf { className == TBigDecimal || className == TBigInteger }
-                ?.takeIf { it.valueArgumentCount == 1 }?.valueArguments?.single()
-                ?.evaluate() ?: return
+            val argVal = expression.takeIf { className.isBigNumber() && it.valueArgumentCount == 1 }
+                ?.valueArguments?.single()?.evaluate() ?: return
 
             when (argVal) {
                 is Number -> constantOfNumber(argVal)
@@ -80,7 +77,7 @@ class BigDecimalConstantInspection : UastInspection(), CleanupLocalInspectionToo
         }
 
         private fun complain(src: PsiElement, prefix: String, replacement: String) {
-            val call = (src as? KtElement)?.getParentOfType<KtDotQualifiedExpression>(true) ?: src
+            val call = (src as? KtElement)?.parent as? KtDotQualifiedExpression ?: src
             val unqualified = prefix.substring(prefix.lastIndexOf('.') + 1) + '.' + replacement
             holder.registerProblem(
                 call,
@@ -115,15 +112,17 @@ class BigDecimalConstantInspection : UastInspection(), CleanupLocalInspectionToo
         }
     }
 
-    companion object {
-        private const val TBigDecimal = "java.math.BigDecimal"
-        private const val TBigInteger = "java.math.BigInteger"
-        private val CONSTRUCTOR_METHOD = CallMatcher.staticCall(TBigDecimal, "valueOf").let {
-            CallMatcher.anyOf(
-                it.parameterTypes("long"),
-                it.parameterTypes("double"),
-                CallMatcher.staticCall(TBigInteger, "valueOf").parameterTypes("long"),
-            )
-        }
-    }
+}
+
+fun String.isBigNumber() = this == TBigDecimal || this == TBigInteger
+
+const val TBigDecimal = "java.math.BigDecimal"
+const val TBigInteger = "java.math.BigInteger"
+
+private val CONSTRUCTOR_METHOD = CallMatcher.staticCall(TBigDecimal, "valueOf").let {
+    CallMatcher.anyOf(
+        it.parameterTypes("long"),
+        it.parameterTypes("double"),
+        CallMatcher.staticCall(TBigInteger, "valueOf").parameterTypes("long"),
+    )
 }